requirements.txt
streamlit>=1.25.0
ollama>=0.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
scikit-learn>=1.2.0
numpy>=1.24.0
tqdm>=4.64.0
# Optional (only if you use embeddings/FAISS later)
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4